unify fl trainer script
